+++ packages/theme/src/cli/utilities/theme-environment/local-assets.ts
